+++ src/types/image.types.ts
@@ -1,0 +1,9 @@
+export type ImageProcessResult = {
+  success: boolean;
+  imageURL?: string;
+  error?: string;
+  captureSuccess?: boolean;
+  uploadSuccess?: boolean;
+  captureError?: string;
+  uploadError?: string;
+};
